Код инициализации БД перенесён в сервис инициализации
