+++ UI/SportsStore/Controllers/DatabaseController.cs
@@ -7,6 +7,7 @@ using Microsoft.Extensions.Logging;
 using SportsStore.DAL.Context;
 using SportsStore.Infrastructure.Extensions;
 using SportsStore.Interfaces.Products;
+using SportsStore.Services.Data;
 
 namespace SportsStore.Controllers
 {
@@ -26,15 +27,15 @@ namespace SportsStore.Controllers
            .View(_Products.Query.OrderBy(p => p.Id).Take(20));
 
         [HttpPost]
-        public async Task<IActionResult> CreateSeedData(int Count, [FromServices] SportStoreDB context)
+        public async Task<IActionResult> CreateSeedData(int Count, [FromServices] SportStoreDBInitializer db)
         {
             _Logger.LogInformation("Инициализция БД данными в количестве {0} элементов", Count);
-            await ClearData(context);
+            await ClearData(db);
             if (Count <= 0)
                 return RedirectToAction(nameof(Index));
 
             var timer = Stopwatch.StartNew();
-            await context.SeedTestDataAsync(Count);
+            await db.SeedTestDataAsync(Count);
             timer.Stop();
             _Logger.LogInformation("Инициализация БД данными в количестве {0} - {1}мс", Count, timer.ElapsedMilliseconds);
 
@@ -42,12 +43,12 @@ namespace SportsStore.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> ClearData([FromServices] SportStoreDB context)
+        public async Task<IActionResult> ClearData([FromServices] SportStoreDBInitializer db)
         {
             _Logger.LogInformation("Запуск очистки БД...");
             var timer = Stopwatch.StartNew();
 
-            await context.ClearDatabaseAsync();
+            await db.ClearDatabaseAsync();
 
             timer.Stop();
             _Logger.LogInformation("Выполнение очистки БД заняло {0} мс", timer.ElapsedMilliseconds);
